Test PR
Our listener does some basic validation and processing of the interactive message payload, and then
dispatches the `photo_details` interactions from our previous attachment to a new function
`handlePhotoDetailsInteraction()`. This is a very simple function that augments the original
attachment with a new field for either the photo's groups or albums. Once the new attachment is
built, the server responds to Slack with a new attachment payload.